+++ src/test/converter2/behavior/excludeCategories.ts
@@ -1,0 +1,14 @@
+/**
+ * @category A
+ */
+export const a = 1;
+/**
+ * @category A
+ * @category B
+ */
+export const b = 1;
+/**
+ * @category C
+ */
+export const c = 1;
+export const d = 1;